acid rain + raw bombing

+++ Source/RimWorldOnlineCity/GameClasses/Incident/OCIncident.cs
@@ -11,7 +11,7 @@ namespace RimWorldOnlineCity
 {
     public abstract class OCIncident
     {
-        public float mult = 1;
+        public int mult = 1;
         public IncidentStrategys strategy = IncidentStrategys.ImmediateAttack;
         public IncidentArrivalModes arrivalMode = IncidentArrivalModes.EdgeWalkIn;
         public string faction = null;
